+++ saleor/graphql/checkout/mutations/checkout_lines_update.py
@@ -9,7 +9,7 @@ from ....webhook.event_types import WebhookEventAsyncType
 from ...app.dataloaders import get_app_promise
 from ...checkout.types import CheckoutLine
 from ...core import ResolveInfo
-from ...core.descriptions import ADDED_IN_321, DEPRECATED_IN_3X_INPUT
+from ...core.descriptions import DEPRECATED_IN_3X_INPUT
 from ...core.doc_category import DOC_CATEGORY_CHECKOUT
 from ...core.scalars import UUID, PositiveDecimal
 from ...core.types import BaseInputObjectType, CheckoutError, NonNullList
@@ -59,7 +59,7 @@ class CheckoutLineUpdateInput(BaseInputObjectType):
     metadata = NonNullList(
         MetadataInput,
         description=(
-            f"Checkout line public metadata. Will add and update keys. To delete keys use deleteMetadata mutation. {ADDED_IN_321} "
+            f"Checkout line public metadata. Will add and update keys. To delete keys use deleteMetadata mutation. "
             f"{MetadataInputDescription.PUBLIC_METADATA_INPUT}"
         ),
         required=False,