+++ saleor/graphql/core/types/common.py
@@ -26,7 +26,6 @@ from ...core.doc_category import (
     DOC_CATEGORY_WEBHOOKS,
 )
 from ...core.scalars import DateTime, Decimal
-from ..descriptions import ADDED_IN_318
 from ..enums import (
     AccountErrorCode,
     AppErrorCode,
@@ -304,7 +303,7 @@ class DiscountError(ProductWithoutVariantError):
     )
     voucher_codes = NonNullList(
         graphene.String,
-        description="List of voucher codes which causes the error." + ADDED_IN_318,
+        description="List of voucher codes which causes the error.",
         required=False,
     )