+++ saleor/graphql/giftcard/mutations/gift_card_create.py
@@ -15,7 +15,7 @@ from ....permission.enums import GiftcardPermissions
 from ....webhook.event_types import WebhookEventAsyncType
 from ...app.dataloaders import get_app_promise
 from ...core import ResolveInfo
-from ...core.descriptions import ADDED_IN_321, DEPRECATED_IN_3X_INPUT
+from ...core.descriptions import DEPRECATED_IN_3X_INPUT
 from ...core.doc_category import DOC_CATEGORY_GIFT_CARDS
 from ...core.mutations import DeprecatedModelMutation
 from ...core.scalars import Date
@@ -37,14 +37,14 @@ class GiftCardInput(BaseInputObjectType):
     metadata = NonNullList(
         MetadataInput,
         description=(
-            f"Gift Card public metadata. {ADDED_IN_321} "
+            f"Gift Card public metadata. "
             f"{MetadataInputDescription.PUBLIC_METADATA_INPUT}"
         ),
         required=False,
     )
     private_metadata = NonNullList(
         MetadataInput,
-        description=f"Gift Card private metadata. {ADDED_IN_321} "
+        description=f"Gift Card private metadata. "
         f"{MetadataInputDescription.PRIVATE_METADATA_INPUT}",
         required=False,
     )