+++ saleor/graphql/order/types.py
@@ -75,10 +75,6 @@ from ..checkout.utils import prevent_sync_event_circular_query
 from ..core.connection import CountableConnection
 from ..core.context import ChannelContext
 from ..core.descriptions import (
-    ADDED_IN_318,
-    ADDED_IN_319,
-    ADDED_IN_320,
-    ADDED_IN_321,
     ADDED_IN_322,
     ADDED_IN_323,
     DEPRECATED_IN_3X_INPUT,
@@ -355,19 +351,17 @@ class OrderGrantedRefund(
         required=True,
         description=(
             "Status of the granted refund calculated based on transactionItem assigned "
-            "to granted refund." + ADDED_IN_320
+            "to granted refund."
         ),
     )
     transaction_events = NonNullList(
         TransactionEvent,
-        description=(
-            "List of refund events associated with the granted refund." + ADDED_IN_320
-        ),
+        description=("List of refund events associated with the granted refund."),
     )
 
     transaction = graphene.Field(
         TransactionItem,
-        description="The transaction assigned to the granted refund." + ADDED_IN_320,
+        description="The transaction assigned to the granted refund.",
     )
 
     class Meta:
@@ -1248,11 +1242,11 @@ class OrderLine(
         required=False, description="Voucher code that was used for this order line."
     )
     is_gift = graphene.Boolean(
-        description="Determine if the line is a gift." + ADDED_IN_319 + PREVIEW_FEATURE,
+        description="Determine if the line is a gift." + PREVIEW_FEATURE,
     )
     discounts = NonNullList(
         "saleor.graphql.discount.types.discounts.OrderLineDiscount",
-        description="List of applied discounts" + ADDED_IN_321,
+        description="List of applied discounts",
     )
 
     class Meta:
@@ -1741,7 +1735,7 @@ class Order(SyncWebhookControlContextModelObjectType[ModelObjectType[models.Orde
     )
     undiscounted_shipping_price = graphene.Field(
         Money,
-        description="Undiscounted total price of shipping." + ADDED_IN_319,
+        description="Undiscounted total price of shipping.",
         required=True,
     )
     shipping_price = graphene.Field(
@@ -1788,7 +1782,7 @@ class Order(SyncWebhookControlContextModelObjectType[ModelObjectType[models.Orde
     voucher = graphene.Field(Voucher, description="Voucher linked to the order.")
     voucher_code = graphene.String(
         required=False,
-        description="Voucher code that was used for Order." + ADDED_IN_318,
+        description="Voucher code that was used for Order.",
     )
     gift_cards = NonNullList(
         GiftCard, description="List of user gift cards.", required=True