+++ saleor/graphql/webhook/subscription_types.py
@@ -50,12 +50,9 @@ from ..core.context import (
     get_database_connection_name,
 )
 from ..core.descriptions import (
-    ADDED_IN_318,
-    ADDED_IN_319,
-    ADDED_IN_320,
-    ADDED_IN_321,
     ADDED_IN_322,
     ADDED_IN_323,
+    ADDED_IN_324,
     DEPRECATED_IN_3X_EVENT,
     PREVIEW_FEATURE,
 )
@@ -902,6 +899,42 @@ class ProductMetadataUpdated(SubscriptionObjectType, ProductBase):
         enable_dry_run = True
         interfaces = (Event,)
         description = "Event sent when product metadata is updated."
+
+
+class ProductTypeBase(AbstractType):
+    product_type = graphene.Field(
+        "saleor.graphql.product.types.ProductType",
+        description="The product type the event relates to.",
+    )
+
+    @staticmethod
+    def resolve_product_type(root, _info: ResolveInfo):
+        _, product_type = root
+        return product_type
+
+
+class ProductTypeCreated(SubscriptionObjectType, ProductTypeBase):
+    class Meta:
+        root_type = "ProductType"
+        enable_dry_run = True
+        interfaces = (Event,)
+        description = "Event sent when new product type is created." + ADDED_IN_324
+
+
+class ProductTypeUpdated(SubscriptionObjectType, ProductTypeBase):
+    class Meta:
+        root_type = "ProductType"
+        enable_dry_run = True
+        interfaces = (Event,)
+        description = "Event sent when product type is updated." + ADDED_IN_324
+
+
+class ProductTypeDeleted(SubscriptionObjectType, ProductTypeBase):
+    class Meta:
+        root_type = "ProductType"
+        enable_dry_run = True
+        interfaces = (Event,)
+        description = "Event sent when product type is deleted." + ADDED_IN_324
 
 
 class ProductMediaBase(AbstractType):
@@ -2534,7 +2567,7 @@ class VoucherCodesCreated(SubscriptionObjectType, VoucherCodeBase):
         root_type = "VoucherCode"
         enable_dry_run = True
         interfaces = (Event,)
-        description = "Event sent when new voucher codes were created." + ADDED_IN_319
+        description = "Event sent when new voucher codes were created."
 
 
 class VoucherCodesDeleted(SubscriptionObjectType, VoucherCodeBase):
@@ -2542,7 +2575,7 @@ class VoucherCodesDeleted(SubscriptionObjectType, VoucherCodeBase):
         root_type = "VoucherCode"
         enable_dry_run = True
         interfaces = (Event,)
-        description = "Event sent when voucher codes were deleted." + ADDED_IN_319
+        description = "Event sent when voucher codes were deleted."
 
 
 class VoucherMetadataUpdated(SubscriptionObjectType, VoucherBase):
@@ -2563,7 +2596,7 @@ class VoucherCodeExportCompleted(SubscriptionObjectType):
         root_type = "ExportFile"
         enable_dry_run = True
         interfaces = (Event,)
-        description = "Event sent when voucher code export is completed." + ADDED_IN_318
+        description = "Event sent when voucher code export is completed."
         doc_category = DOC_CATEGORY_DISCOUNTS
 
     @staticmethod
@@ -2826,56 +2859,42 @@ class Subscription(SubscriptionObjectType):
     )
     draft_order_created = BaseField(
         DraftOrderCreated,
-        description=(
-            "Event sent when new draft order is created."
-            + ADDED_IN_320
-            + PREVIEW_FEATURE
-        ),
+        description=("Event sent when new draft order is created." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     draft_order_updated = BaseField(
         DraftOrderUpdated,
-        description=(
-            "Event sent when draft order is updated." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when draft order is updated." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     draft_order_deleted = BaseField(
         DraftOrderDeleted,
-        description=(
-            "Event sent when draft order is deleted." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when draft order is deleted." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_created = BaseField(
         OrderCreated,
-        description=(
-            "Event sent when new order is created." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when new order is created." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_updated = BaseField(
         OrderUpdated,
-        description=(
-            "Event sent when order is updated." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when order is updated." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_confirmed = BaseField(
         OrderConfirmed,
-        description=(
-            "Event sent when order is confirmed." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when order is confirmed." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
@@ -2884,7 +2903,6 @@ class Subscription(SubscriptionObjectType):
         OrderPaid,
         description=(
             "Payment has been made. The order may be partially or fully paid."
-            + ADDED_IN_320
             + PREVIEW_FEATURE
         ),
         resolver=default_channel_filterable_resolver,
@@ -2893,9 +2911,7 @@ class Subscription(SubscriptionObjectType):
     )
     order_fully_paid = BaseField(
         OrderFullyPaid,
-        description=(
-            "Event sent when order is fully paid." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when order is fully paid." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
@@ -2904,7 +2920,7 @@ class Subscription(SubscriptionObjectType):
         OrderRefunded,
         description=(
             "The order received a refund. The order may be partially or fully "
-            "refunded." + ADDED_IN_320 + PREVIEW_FEATURE
+            "refunded." + PREVIEW_FEATURE
         ),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
@@ -2912,103 +2928,77 @@ class Subscription(SubscriptionObjectType):
     )
     order_fully_refunded = BaseField(
         OrderFullyRefunded,
-        description=("The order is fully refunded." + ADDED_IN_320 + PREVIEW_FEATURE),
+        description=("The order is fully refunded." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_fulfilled = BaseField(
         OrderFulfilled,
-        description=(
-            "Event sent when order is fulfilled." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when order is fulfilled." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_cancelled = BaseField(
         OrderCancelled,
-        description=(
-            "Event sent when order is cancelled." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when order is cancelled." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_expired = BaseField(
         OrderExpired,
-        description=(
-            "Event sent when order becomes expired." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when order becomes expired." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_metadata_updated = BaseField(
         OrderMetadataUpdated,
-        description=(
-            "Event sent when order metadata is updated."
-            + ADDED_IN_320
-            + PREVIEW_FEATURE
-        ),
+        description=("Event sent when order metadata is updated." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
     order_bulk_created = BaseField(
         OrderBulkCreated,
-        description=(
-            "Event sent when orders are imported." + ADDED_IN_320 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when orders are imported." + PREVIEW_FEATURE),
         channels=channels_argument,
         doc_category=DOC_CATEGORY_ORDERS,
     )
 
     checkout_created = BaseField(
         CheckoutCreated,
-        description=(
-            "Event sent when new checkout is created." + ADDED_IN_321 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when new checkout is created." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_CHECKOUT,
     )
     checkout_updated = BaseField(
         CheckoutUpdated,
-        description=(
-            "Event sent when checkout is updated." + ADDED_IN_321 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when checkout is updated." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_CHECKOUT,
     )
     checkout_fully_paid = BaseField(
         CheckoutFullyPaid,
-        description=(
-            "Event sent when checkout is fully-paid." + ADDED_IN_321 + PREVIEW_FEATURE
-        ),
+        description=("Event sent when checkout is fully-paid." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_CHECKOUT,
     )
     checkout_fully_authorized = BaseField(
         CheckoutFullyAuthorized,
-        description=(
-            "Event sent when checkout is fully authorized."
-            + ADDED_IN_321
-            + PREVIEW_FEATURE
-        ),
+        description=("Event sent when checkout is fully authorized." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_CHECKOUT,
     )
     checkout_metadata_updated = BaseField(
         CheckoutMetadataUpdated,
-        description=(
-            "Event sent when checkout metadata is updated."
-            + ADDED_IN_321
-            + PREVIEW_FEATURE
-        ),
+        description=("Event sent when checkout metadata is updated." + PREVIEW_FEATURE),
         resolver=default_channel_filterable_resolver,
         channels=channels_argument,
         doc_category=DOC_CATEGORY_CHECKOUT,
@@ -3244,6 +3234,9 @@ ASYNC_WEBHOOK_TYPES_MAP = {
     WebhookEventAsyncType.PRODUCT_DELETED: ProductDeleted,
     WebhookEventAsyncType.PRODUCT_METADATA_UPDATED: ProductMetadataUpdated,
     WebhookEventAsyncType.PRODUCT_EXPORT_COMPLETED: ProductExportCompleted,
+    WebhookEventAsyncType.PRODUCT_TYPE_CREATED: ProductTypeCreated,
+    WebhookEventAsyncType.PRODUCT_TYPE_UPDATED: ProductTypeUpdated,
+    WebhookEventAsyncType.PRODUCT_TYPE_DELETED: ProductTypeDeleted,
     WebhookEventAsyncType.PRODUCT_MEDIA_CREATED: ProductMediaCreated,
     WebhookEventAsyncType.PRODUCT_MEDIA_UPDATED: ProductMediaUpdated,
     WebhookEventAsyncType.PRODUCT_MEDIA_DELETED: ProductMediaDeleted,